Add mulx inline assembly code in comment
The intrinsic core::arch::x86_64::_mulx_u64 doesn't work.
_mulx_u64 relies on LLVM to lower a u128 multiplication to mulx which it doesn't do anymore on 1.34.0 (see #1).
As we don't want to rely on nightly, I comment this code for now.

But clearly this intrinsic should not be in core::arch and not advertising it is not affecting flags.
I propose to the Compiler team to remove it since it is equivalent to a u128 multiplication.